requirements.txt
amqp==5.0.6
asgiref==3.4.1
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
coverage==6.0
cryptography==3.4.7
decorator==5.0.9
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.5
django-admin-rangefilter==0.8.1
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.1
django-extensions==3.1.3
django-filter==2.4.0
django-import-export==2.5.0
django-silk==4.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-yasg==1.20.0
et-xmlfile==1.1.0
flake8==3.9.2
flake8-builtins==1.5.3
flake8-import-order==0.18.1
flake8-print==4.0.0
gprof2dot==2021.2.21
idna==3.2
importlib-metadata==4.6.1
inflection==0.5.1
iniconfig==1.1.1
ipython==7.25.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
kombu==5.1.0
MarkupPy==1.14
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.7
packaging==21.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.0.0
prompt-toolkit==3.0.19
psycopg2-binary==2.9.2
ptyprocess==0.7.0
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.9.0
PyJWT==2.1.0
pymemcache==3.5.0
pyparsing==2.4.7
pytest==6.2.5
pytest-cov==2.12.1
pytest-django==4.4.0
pytest-mock==3.6.1
pytest-socket==0.4.1
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
tomli==1.2.1
traitlets==5.0.5
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
uWSGI==2.0.20
vine==5.0.0
wcwidth==0.2.5
xlrd==2.0.1
xlwt==1.3.0
zipp==3.5.0
humanize==3.13.1
flower==1.0.0
django-environ==0.8.1 